requirements-dev.txt
numpy
matplotlib
geopandas>=0.12.0
scikit-learn
networkx 
geoalchemy2 
osmnx 
psycopg2
tqdm
similaritymeasures
# additional dependencies for development
black   # linting
jupyter # notebooks
pytest  # testing
sphinx  # documentation
sphinx_rtd_theme